Extracting types and duplicated function.
